moved all search related stuff to a folder
