+++ react-app/src/components/SupportPage/SearchSupport/searchBarUI.tsx
