+++ react-app/src/components/SupportPage/index.ts
@@ -1,8 +1,8 @@
 import ContactHelpDesk from "./contactHelpDesk";
 import ExpandCollapseBtn from "./expandCollapseBtn";
 import LeftNavigation from "./navigationBar";
-import Search from "./search";
-import SearchContent from "./searchSupport";
+import Search from "./SearchSupport/searchBarUI";
+import SearchContent from "./SearchSupport/searchSupportPage";
 import StatusLabel from "./statusLabel";
 
 export { ContactHelpDesk, ExpandCollapseBtn, LeftNavigation, StatusLabel, Search, SearchContent };
